pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "python-dotenv",
]
